+++ src/molviewspec/mod.rs
@@ -1,3 +1,1 @@
-pub mod builder;
 pub mod nodes;
-// pub mod open_api;
